Добавил модуль 'Реестр выплаченных выигрышей'
